chore: moved cli into ts file under source folder
