+++ packages/compiler/.bin/morfeo.js
@@ -1,72 +1,3 @@
 #!/usr/bin/env node
 
-const { cosmiconfig } = require('cosmiconfig');
-const { TypeScriptLoader } = require('cosmiconfig-typescript-loader');
-const { Command, Option } = require('commander');
-const { collector } = require('../commonjs/collector');
-const { logger } = require('../commonjs/logger');
-
-const moduleName = 'morfeo';
-
-const explorer = cosmiconfig(moduleName, {
-  searchPlaces: [
-    `.${moduleName}rc.js`,
-    `.${moduleName}rc.ts`,
-    `.${moduleName}rc.cjs`,
-    `${moduleName}.config.js`,
-    `${moduleName}.config.ts`,
-    `${moduleName}.config.cjs`,
-  ],
-  loaders: {
-    '.ts': TypeScriptLoader(),
-  },
-});
-
-const program = new Command();
-
-program
-  .name('morfeo')
-  .version('1.0.0', '-v, --version')
-  .description("Run the morfeo's CLI to extract the CSS")
-  .option(
-    '-e, --entryPoints',
-    'Entry points where morfeo should collect the files',
-  )
-  .option('-c, --classNamePrefix', 'The prefix for each generated class')
-  .option('-o, --output', 'The path of the file where the css will be stored')
-  .option(
-    '-E, --emojis',
-    'Whether or not the minified classes should use emojis 😎',
-  )
-  .option('-m, --minify', 'Whether or not the classes should be minified')
-  .option(
-    '-w, --watch [boolean]',
-    'Indicated if morfeo should watch for file changes',
-  )
-  .parse(process.argv);
-
-const configFromCli = program.opts();
-
-function runMorfeo(config) {
-  collector.init({
-    ...config,
-    ...configFromCli,
-  });
-  collector.collect();
-}
-
-function onError() {
-  logger.warning('morfeo configuration file not found.');
-
-  runMorfeo(configFromCli);
-}
-
-function onFound(result) {
-  if (!result) {
-    return onError();
-  }
-  logger.success(`morfeo configuration file found at ${result.filepath}`);
-  runMorfeo(result.config);
-}
-
-explorer.search().then(onFound).catch(onError);
+require('../commonjs/cli');
